requirements.txt
tensorflow-cpu
pandas
streamlit
pillow
matplotlit